put together all primitive serializers
